support deleting substreams from superstreams. removed debug logging from SSTag.

+++ client/models/SSTag.js
@@ -92,14 +92,9 @@ var SSTag = new Class({
     {
       // this is a category tag (superstream), find the real event id by the stream id of the substream. - David
       this.coll().read(function(events) {
-        SSLog('events', SSLogForce);
-        SSLog(events, SSLogForce);
         window.dbg = events;
         var eventToDelete = events.select(function(v) { return v.object_ref == id; });
-        SSLog('eventToDelete', SSLogForce);
-        SSLog(eventToDelete, SSLogForce);
-        this.deleteEvent(eventToDelete.id);
-        SSLog('done', SSLogForce);
+        if(eventToDelete) this.deleteEvent(eventToDelete.id);
       }.bind(this), {bare:1});
     }
   },
